Placement: Use clockwise order for concept maps
